requirements.txt
# Core Bot Framework
python-telegram-bot[job-queue]==20.7
pyrogram==2.0.106
tgcrypto==1.2.5

# HTTP & Networking
httpx[http2]==0.27.0
aiohttp==3.9.4
aiohttp-cors==0.7.0

# HTML Parsing
selectolax==0.4.6
beautifulsoup4==4.12.3
lxml==5.2.1

# Utilities
psutil==5.9.8
python-dotenv==1.0.0
Pillow==10.2.0

# Development
black==23.12.1
pytest==7.4.4
pytest-asyncio==0.21.1

# Optional (for subtitle support)
requests==2.31.0
chardet==5.2.0